Move instrument code under it's own package
