+++ pkg/instrument/instrument.go
@@ -1,4 +1,4 @@
-package aws
+package instrument
 
 import (
 	"errors"
